[rotated-quant] merge-fix: forward state_dtype as c4/c128_state_dtype to super
Merging upupstream/main split DeepSeekV4TokenToKVPool.__init__'s single
`state_dtype` into `c4_state_dtype` + `c128_state_dtype` (and added
sliding_window / online_mtp_max_draft_tokens / num_req_slots, all with
defaults). The fork's RotatedQuantDeepSeekV4TokenToKVPool still declared
and forwarded a single `state_dtype`, so super().__init__ rejected it:
  TypeError: DeepSeekV4TokenToKVPool.__init__() got an unexpected
  keyword argument 'state_dtype'

_get_dsv4_compress_state_dtypes() always returns the SAME dtype for c4
and c128, and the configurator's rotated-path adapter
(kv_cache_configurator.py:1000) already collapses them back into one
`state_dtype` before calling us — so forward it as both c4_state_dtype
and c128_state_dtype. Semantically equivalent to the pre-merge single-
dtype behavior. Server was crashing at pool construction (all 8 TP
ranks) before this.

+++ python/sglang/srt/mem_cache/rotated_quant_dsv4_memory_pool.py
@@ -541,7 +541,13 @@ class RotatedQuantDeepSeekV4TokenToKVPool(DeepSeekV4TokenToKVPool):
             page_size=page_size,
             swa_page_size=swa_page_size,
             dtype=dtype,
-            state_dtype=state_dtype,
+            # [kvbit merge-fix] upstream split the single state_dtype into
+            # c4_state_dtype + c128_state_dtype; _get_dsv4_compress_state_dtypes
+            # always returns the same dtype for both, and the configurator's
+            # rotated-path adapter (kv_cache_configurator.py:1000) collapses them
+            # back into one `state_dtype` before calling us, so forward it as both.
+            c4_state_dtype=state_dtype,
+            c128_state_dtype=state_dtype,
             qk_nope_head_dim=qk_nope_head_dim,
             qk_rope_head_dim=qk_rope_head_dim,
             indexer_head_dim=indexer_head_dim,
